evaluate-candidates/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
